Callback wrapper update
