+++ CHANGELOG.md
@@ -1,19 +1,25 @@
+## 0.6.0
+
+* Callback enabling process simplified:
+    * Kotlin PipCallbackHelperActivityWrapper added
+    * Example app updated to show the new wrapper usage
+
 ## 0.5.1
 
 * Initial release bugfix:
-  * README fixed
-  * SDK min version fixed
+    * README fixed
+    * SDK min version fixed
 
 ## 0.5.0
 
 * Initial development release:
-  * SimplePip class added with features:
-    * isPipAvailable
-    * isPipActivated
-    * enterPipMode
-    * callbacks
-  * PipWidget widget added with features:
-    * builder
-    * pipBuilder
-    * callbacks
-  * Kotlin PipCallbackHelper class added
+    * SimplePip class added with features:
+        * isPipAvailable
+        * isPipActivated
+        * enterPipMode
+        * callbacks
+    * PipWidget widget added with features:
+        * builder
+        * pipBuilder
+        * callbacks
+    * Kotlin PipCallbackHelper class added
