+++ android/src/main/kotlin/cl/puntito/simple_pip_mode/PipCallbackHelperActivityWrapper.kt
@@ -1,0 +1,22 @@
+package cl.puntito.simple_pip_mode
+
+import android.content.res.Configuration
+import androidx.annotation.NonNull
+import io.flutter.embedding.android.FlutterActivity
+import io.flutter.embedding.engine.FlutterEngine
+import cl.puntito.simple_pip_mode.PipCallbackHelper
+
+
+open class PipCallbackHelperActivityWrapper: FlutterActivity() {
+  private var callbackHelper = PipCallbackHelper()
+
+  override fun configureFlutterEngine(@NonNull flutterEngine: FlutterEngine) {
+    super.configureFlutterEngine(flutterEngine)
+    callbackHelper.configureFlutterEngine(flutterEngine)
+  }
+
+  override fun onPictureInPictureModeChanged(active: Boolean, newConfig: Configuration?) {
+    super.onPictureInPictureModeChanged(active, newConfig)
+    callbackHelper.onPictureInPictureModeChanged(active)
+  }
+}
